CHG: Position buttons
vertical position: For the submit button we be use the bottom of the current answer text field, but for the clear button we be set its Y anchor so that its stays aligned with the Y position of the submit button. This means both buttons will remain aligned even if we move one.